+++ src/cli/commands/install.rs
@@ -10,13 +10,14 @@ use crate::{
             logging::{error, warning},
             not_found,
         },
+        parameter_checks,
     },
     config::Config,
     installer::{InstallType, Installer, InstallerOptions, types::OptionalPackageId},
     platforms::Target,
     register::package_register::PackageRegister,
     repositories::{error::RepositoryError, manager::RepositoryManager, types::Date},
-    utils::{duplicates, unwrap_or_exit::UnwrapOrExit},
+    utils::unwrap_or_exit::UnwrapOrExit,
 };
 
 /// Installs the specified packages, if a version is given that version will be installed,
@@ -56,7 +57,7 @@ pub struct InstallArgs {
 impl HandleCommand for InstallArgs {
     fn handle(&self) {
         // Check for duplicates, because installing twice will result in a confusing error
-        let duplicates = duplicates::get_duplicates(&self.packages);
+        let duplicates = parameter_checks::get_duplicates(&self.packages);
         if !duplicates.is_empty() {
             let mut duplicate_string = String::new();
             for duplicate in duplicates {